requirements.txt
pillow>=11.2.1
